use hypothesis strategy to test load(dump(g))
Also closes #136.
