+++ tests/test_demes.py
@@ -7,7 +7,6 @@ import math
 import jsonschema
 import pytest
 import hypothesis as hyp
-import hypothesis.strategies as st
 
 from demes import (
     Epoch,
@@ -21,220 +20,27 @@ from demes import (
     Admix,
 )
 import demes
-
-
-@st.composite
-def epochs_lists(draw, start_time=math.inf, max_epochs=10):
-    """
-    A hypothesis strategy for creating lists of Epochs for a deme.
-
-    .. code-block::
-
-        @hypothesis.given(epoch_lists())
-        test_something(self, epoch_list):
-            # epoch_list has type ``list of Epoch``
-            pass
-
-    :param start_time: the start time of the deme.
-    :param max_epochs: the maximum number of epochs in the list.
-    """
-    times = draw(
-        st.sets(
-            st.floats(min_value=0, max_value=start_time),
-            min_size=2,
-            max_size=max_epochs,
-        )
-    )
-    times = sorted(list(times), reverse=True)
-    if start_time != times[0]:
-        times.insert(0, start_time)
-    epochs = []
-
-    for end_time in times[1:]:
-        initial_size = draw(
-            st.floats(min_value=0, exclude_min=True, allow_infinity=False)
-        )
-        if math.isinf(start_time):
-            final_size = initial_size
-        else:
-            final_size = draw(
-                st.floats(min_value=0, exclude_min=True, allow_infinity=False)
-            )
-        cloning_rate = draw(st.floats(min_value=0, max_value=1))
-        selfing_rate = draw(st.floats(min_value=0, max_value=1))
-
-        epochs.append(
-            Epoch(
-                start_time=start_time,
-                end_time=end_time,
-                initial_size=initial_size,
-                final_size=final_size,
-                cloning_rate=cloning_rate,
-                selfing_rate=selfing_rate,
-            )
-        )
-        start_time = end_time
-
-    return epochs
-
-
-@st.composite
-def graphs(draw, max_demes=10, max_interactions=10):
-    """
-    A hypothesis strategy for create a Graph.
-
-    .. code-block::
-
-        @hypothesis.given(graphs())
-        def test_something(self, g):
-            # g has type ``Graph``
-            pass
-
-    :param max_demes: The maximum number of demes in the graph.
-    :param max_interactions: The maximum number of migrations plus pulses
-        in the graph.
-    """
-    generation_time = draw(st.none() | st.floats(min_value=1e-9, max_value=1e6))
-    if generation_time is None:
-        time_units = "generations"
-    else:
-        time_units = draw(st.text(max_size=100))
-    g = Graph(
-        description=draw(st.text(max_size=100)),
-        generation_time=generation_time,
-        time_units=time_units,
-        doi=draw(st.lists(st.text(min_size=1), max_size=3)),
-    )
-    deme_ids = draw(st.sets(st.text(min_size=1), min_size=1, max_size=max_demes))
-
-    for id in deme_ids:
-        ancestors = []
-        proportions = []
-        start_time = math.inf
-        if len(g.demes) > 0:
-            # draw indices into demes list to use as ancestors
-            anc_idx = draw(
-                st.lists(
-                    st.integers(min_value=0, max_value=len(g.demes) - 1),
-                    unique=True,
-                    max_size=len(g.demes),
-                )
-            )
-            if len(anc_idx) > 0:
-                time_hi = min(g.demes[j].start_time for j in anc_idx)
-                time_lo = max(g.demes[j].end_time for j in anc_idx)
-                if time_lo < time_hi and time_lo < 1e308:
-                    # The proposed ancestors exist at the same time.
-                    # Draw a start time and the ancestry proportions.
-                    start_time = draw(
-                        st.floats(
-                            min_value=time_lo,
-                            max_value=time_hi,
-                            exclude_max=True,
-                        )
-                    )
-                    ancestors = [g.demes[j].id for j in anc_idx]
-                    if len(ancestors) == 1:
-                        proportions = [1.0]
-                    else:
-                        proportions = draw(
-                            st.lists(
-                                st.integers(min_value=1, max_value=1000),
-                                min_size=len(ancestors),
-                                max_size=len(ancestors),
-                            )
-                        )
-                        psum = sum(proportions)
-                        proportions = [p / psum for p in proportions]
-        g.deme(
-            id=id,
-            description=draw(st.none() | st.text(max_size=100)),
-            ancestors=ancestors,
-            proportions=proportions,
-            epochs=draw(epochs_lists(start_time=start_time)),
-            start_time=start_time,
-        )
-
-    n_interactions = draw(st.integers(min_value=0, max_value=max_interactions))
-    for j in range(len(g.demes) - 1):
-        for k in range(j + 1, len(g.demes)):
-            dj = g.demes[j].id
-            dk = g.demes[k].id
-            time_lo = max(g[dj].end_time, g[dk].end_time)
-            time_hi = min(g[dj].start_time, g[dk].start_time)
-            if time_hi <= time_lo or time_lo > 1e308:
-                # Demes j and k don't exist at the same time.
-                # (or time_lo is too close to infinity for floats)
-                continue
-            # Draw asymmetric migrations.
-            n = draw(st.integers(min_value=0, max_value=n_interactions))
-            n_interactions -= n
-            for _ in range(n):
-                source, dest = dj, dk
-                if draw(st.booleans()):
-                    source, dest = dk, dj
-                times = draw(
-                    st.lists(
-                        st.floats(min_value=time_lo, max_value=time_hi),
-                        unique=True,
-                        min_size=2,
-                        max_size=2,
-                    )
-                )
-                g.migration(
-                    source=source,
-                    dest=dest,
-                    start_time=max(times),
-                    end_time=min(times),
-                    rate=draw(st.floats(min_value=0, max_value=1, exclude_max=True)),
-                )
-            if n_interactions <= 0:
-                break
-            # Draw pulses.
-            n = draw(st.integers(min_value=0, max_value=n_interactions))
-            n_interactions -= n
-            for _ in range(n):
-                source, dest = dj, dk
-                if draw(st.booleans()):
-                    source, dest = dk, dj
-                time = draw(
-                    st.floats(
-                        min_value=time_lo,
-                        max_value=time_hi,
-                        exclude_min=True,
-                        exclude_max=True,
-                    )
-                )
-                g.pulse(
-                    source=source,
-                    dest=dest,
-                    time=time,
-                    proportion=draw(
-                        st.floats(
-                            min_value=0, max_value=1, exclude_min=True, exclude_max=True
-                        )
-                    ),
-                )
-            if n_interactions <= 0:
-                break
-        if n_interactions <= 0:
-            break
-
-    return g
+import tests
 
 
 class TestHypothesisStrategies:
     # Test that the hypothesis strategies can be used.
 
-    @hyp.given(epochs_lists())
-    def test_epoch_construction(self, e):
+    @hyp.given(tests.epochs_lists())
+    def test_epochs_lists(self, e):
         # e is a list of Epoch
         pass
 
-    @hyp.given(graphs())
-    def test_graph_construction(self, g):
+    @hyp.settings(suppress_health_check=[hyp.HealthCheck.too_slow])
+    @hyp.given(tests.graphs())
+    def test_graphs(self, g):
         # g is a Graph
         g.validate()
+
+    @hyp.given(tests.deme_ids())
+    def test_deme_ids(self, id):
+        # id is a valid deme ID
+        id.isidentifier()
 
 
 class TestEpoch(unittest.TestCase):
@@ -795,14 +601,15 @@ class TestDeme(unittest.TestCase):
                 proportions=[],
                 epochs=[Epoch(start_time=math.inf, end_time=0, initial_size=1)],
             )
-        with self.assertRaises(ValueError):
-            Deme(
-                id="",
-                description="b",
-                ancestors=[],
-                proportions=[],
-                epochs=[Epoch(start_time=math.inf, end_time=0, initial_size=1)],
-            )
+        for bad_id in ["", "501", "pop-1", "pop.2"]:
+            with self.assertRaises(ValueError):
+                Deme(
+                    id=bad_id,
+                    description="b",
+                    ancestors=[],
+                    proportions=[],
+                    epochs=[Epoch(start_time=math.inf, end_time=0, initial_size=1)],
+                )
 
     def test_bad_ancestors(self):
         with self.assertRaises(TypeError):
@@ -1079,13 +886,25 @@ class TestDeme(unittest.TestCase):
 
 class TestGraph(unittest.TestCase):
     def test_bad_generation_time(self):
-        for generation_time in (-100, -1e-9, 0, float("inf")):
+        for generation_time in (-100, -1e-9, 0, float("inf"), None):
             with self.assertRaises(ValueError):
                 Graph(
                     description="test",
                     time_units="years",
                     generation_time=generation_time,
                 )
+
+    def test_bad_description(self):
+        with self.assertRaises(TypeError):
+            Graph(
+                description=None,
+                time_units="generations",
+            )
+        with self.assertRaises(ValueError):
+            Graph(
+                description="",
+                time_units="generations",
+            )
 
     def test_doi(self):
         # We currently accept arbitrary strings in DOIs.
@@ -1137,7 +956,8 @@ class TestGraph(unittest.TestCase):
                 doi="10.1000/123456",
             )
 
-    @hyp.given(graphs(max_demes=5, max_interactions=5))
+    @hyp.settings(suppress_health_check=[hyp.HealthCheck.too_slow])
+    @hyp.given(tests.graphs())
     def test_in_generations(self, dg1):
         if dg1.generation_time is None:
             dg1.time_units = "years"
